% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{arsatfc}
\alias{arsatfc}
\title{Ar Saturation}
\usage{
arsatfc(temp, salinity, bp)
}
\arguments{
\item{temp}{Temperature (Celcius)}

\item{salinity}{Salinity}

\item{bp}{Barymetric pressure (elevation-corrected mm Hg)}
}
\description{
Function to calculate the ar saturation.
}
\examples{

arsatfc(15, 0, 760)

}
