% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_dataList}
\alias{create_dataList}
\title{Data list}
\usage{
create_dataList(data, Kmean = 4.03, Ksd = 4, up = "up1",
  down = "down1", tt = 0.1909720833, z = 0.5588)
}
\arguments{
\item{data}{}
}
\description{
Data list
}
