% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fitmod}
\alias{fitmod}
\title{Estimate}
\usage{
fitmod(data, model = 3, chains, verbose = FALSE, warmup)
}
\arguments{
\item{data}{data.frame object with columnds for station, temperature, n2.ar, light, bp, z, and dtime, arsat, and n2sat}

\item{model}{Model number for fitting algorithm.}
}
\description{
Estimate
}
\details{
Model determines which model to estimate. 3 being the two-station model without N consumption (DN base), and 4 being the two station model with N consumption (DN N consume).
}
