% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_likelihood.R
\name{model_likelihood}
\alias{model_likelihood}
\title{Outpout model likelihood}
\usage{
model_likelihood(dataList, paramsVec, model, incl_DN = TRUE)
}
\arguments{
\item{dataList}{dataList created by \code{create_dataList}}

\item{paramsVec}{a parameter vector of length 4 or 5 created by the first column output by \code{plotmod}. If model == 1 or 3, the order will be: DN, K600, sigma2, logpost. If model == 2 or 4 will be: K600, DN, Nfix, sigma2, logpost.}

\item{model}{Model number for fitting algorithm (see details bleow).}
}
\description{
Function to output the model likelihhood given a dataset, model, and vector of parameters. Note this does not include the prior.
}
\details{
Model determines which model to estimate. 0 is the oxygen model (Eq. 2 from Nifong et al), 1 is the single station model without N consumption (DN base model Eq. 3 from Nifong et al.), 2 is the single station model with N consumption (DN + Nconsume Eq. 4 from Nifong et al.), the  3 being the two-station model without N consumption (DN base Eq. 5 from Nifong et al.), and 4 being the two station model with N consumption (DN N consume Eq. 6 from Nifong et al.).
}
