% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{n2satfc}
\alias{n2satfc}
\title{N2 saturation}
\usage{
n2satfc(temp, salinity, bp)
}
\arguments{
\item{temp}{Temperature (Celcius)}

\item{salinity}{Salinity}

\item{bp}{Barymetric pressure (elevation-corrected mm Hg)}
}
\description{
Function to calculate the N2 saturation.
}
\examples{

n2satfc(15, 0, 760)

}
