% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{osat}
\alias{osat}
\title{Oxygen saturation}
\usage{
osat(temp, bp)
}
\arguments{
\item{temp}{Water temperature (Celcius)}

\item{bp}{Barymetric pressure (elevation-corrected mm Hg)}
}
\description{
From Garcia and Gordon 1992 L&O. This calculation is a slight approximation because we do not account for water density, but the difference is much smaller (0.01 mg/L at 20 deg) than anyone's ability to calibrate an oxygen sonde.
}
\examples{

osat(26, 0)
}
