% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotmod}
\alias{plotmod}
\title{Plot and diagnost}
\usage{
plotmod(StanFit, dataList, model = 3, file = NULL)
}
\arguments{
\item{StanFit}{Standmodel output by \code{fitmod}}

\item{dataList}{dataList created by \code{create_dataList}}

\item{model}{Model number for fitting algorithm.}

\item{file}{filname to save the parameter estimates. Will not save if NULL.}
}
\description{
Plot and diagnost
}
\details{
Model determines which model to estimate. 3 being the two-station model without N consumption (DN base), and 4 being the two station model with N consumption (DN N consume).
}
