% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{watdens}
\alias{watdens}
\title{Water density}
\usage{
watdens(temp)
}
\arguments{
\item{temp}{Water temperature (Celcius)}
}
\description{
Function to calculate the water density as a function of temperature.
}
\examples{

watdens(29)

}
