% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_dataList}
\alias{create_dataList}
\title{Data list}
\usage{
create_dataList(data, model = 1, Kmean = 4.03, Ksd = 4, up = "up1",
  down = "down1", tt = 0.1909720833, depth = 0.5588,
  PPFDstart = 14)
}
\arguments{
\item{data}{data.frame object with columnds for station, temperature, n2.ar, light, bp, z, and dtime, arsat, and n2sat}

\item{model}{Model number for fitting algorithm.}

\item{Kmean}{Mean of normal prior distribution for K600}

\item{Ksd}{Sd of normal prior distribution for K600}

\item{up}{Name indicating the up-river station name}

\item{down}{Name indicating the down-river station name}

\item{tt}{Time between stations}

\item{depth}{Depth (m)}

\item{PPFDstart}{Start time (hours) for calculating daily total of photosynthetic photon flux density (PPFD). Calculated by summing across column light for the next 24 hours from the first PPFDstart time.}
}
\description{
Creates a list of data objects to be used by stan
}
\details{
Model determines which model to estimate. 1 is the single station model without N consumption (DN base model; Eq. 3 from Nifong et al.), 2 is the single station model with N consumption (DN + Nconsume; Eq. 4 from Nifong et al.), the  3 being the two-station model without N consumption (DN base; Eq. 5 from Nifong et al.), and 4 being the two station model with N consumption (DN N consume; Eq. 6 from Nifong et al.).
}
