% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fitmod}
\alias{fitmod}
\title{Estimate model}
\usage{
fitmod(dataList, model = 3, nChains = 2, niter = 5000,
  burnin = 1000, verbose = FALSE, incl_DN = TRUE, control = NULL)
}
\arguments{
\item{dataList}{dataList created by \code{create_dataList}}

\item{model}{Model number for fitting algorithm (see details bleow).}

\item{nChains}{Number of chains to run}

\item{niter}{Number of iterations to run for each chain (including burnin)}

\item{burnin}{A positive integer specifying the number of warmup (aka burnin) iterations per chain.}

\item{verbose}{TRUE or FALSE: flag indicating whether to print intermediate output from Stan on the console, which might be helpful for model debugging.}

\item{control}{A named \code{list} of parameters to control the sampler's behavior. See the details in the documentation for the \code{control} argument in \code{?rstan::stan}.}
}
\description{
fits a stan model and returns a stan model object
}
\details{
Model determines which model to estimate. 0 is the oxygen model (Eq. 2 from Nifong et al), 1 is the single station model without N consumption (DN base model; Eq. 3 from Nifong et al.), 2 is the single station model with N consumption (DN + Nconsume; Eq. 4 from Nifong et al.), the  3 being the two-station model without N consumption (DN base; Eq. 5 from Nifong et al.), and 4 being the two station model with N consumption (DN N consume; Eq. 6 from Nifong et al.).
}
\examples{

# Run model Eq. 5
data(InitialData)
dataList <- create_dataList(InitialData , Kmean = 4.03, Ksd = 4.0, up = "up1", down = "down1", tt = 0.1909720833, depth = 0.5588)
mod <- fitmod(dataList, model = 3)
plotmod(mod, dataList = dataList, model = 3, file = NULL)

# Run model Eq. 6
data(InitialData)
dataList <- create_dataList(InitialData , Kmean = 4.03, Ksd = 4.0, up = "up1", down = "down1", tt = 0.19097290833, depth = 0.5588)
mod2 <- fitmod(dataList, model = 4, verbose = FALSE)
plotmod(StanFit = mod2, dataList = dataList, model = 4)

}
