% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plotmod}
\alias{plotmod}
\title{Plot and diagnose}
\usage{
plotmod(StanFit, dataList, model = 3, file = NULL, incl_DN = TRUE)
}
\arguments{
\item{StanFit}{Standmodel output by \code{fitmod}}

\item{dataList}{dataList created by \code{create_dataList}}

\item{model}{Model number for fitting algorithm (see details bleow).}

\item{file}{filname to save the parameter estimates. Will not save if NULL.}
}
\description{
Plot and diagnose
}
\details{
Model determines which model to estimate. 0 is the oxygen model (Eq. 2 from Nifong et al), 1 is the single station model without N consumption (DN base model; Eq. 3 from Nifong et al.), 2 is the single station model with N consumption (DN + Nconsume; Eq. 4 from Nifong et al.), the  3 being the two-station model without N consumption (DN base; Eq. 5 from Nifong et al.), and 4 being the two station model with N consumption (DN N consume; Eq. 6 from Nifong et al.).
}
\examples{
# Run model Eq. 5 from Nifong et al.
data(InitialData)
dataList <- create_dataList(InitialData , Kmean = 4.03, Ksd = 4.0, up = "up1", down = "down1", tt = 0.1909720833, depth = 0.5588)
mod <- fitmod(dataList, model = 3)
plotmod(mod, dataList = dataList, model = 3, file = NULL)

# Run model Eq. 6 from Nifong et al.
data(InitialData)
dataList <- create_dataList(InitialData , Kmean = 4.03, Ksd = 4.0, up = "up1", down = "down1", tt = 0.19097290833, depth = 0.5588)
mod2 <- fitmod(dataList, model = 4, verbose = FALSE)
plotmod(StanFit = mod2, dataList = dataList, model = 4)

}
